// src/schemas/feedback.schema.ts
import { z } from 'zod';

export const FeedbackSchema = z.object({
    id: z.uuid().optional(), // auto-generated by Prisma
    name: z
        .string()
        .min(1, 'Name is required')
        .max(100, 'Name cannot exceed 100 characters'),
    designation: z
        .string()
        .min(1, 'Designation is required')
        .max(100, 'Designation cannot exceed 100 characters'),
    comment: z
        .string()
        .min(1, 'Comment is required')
        .max(500, 'Comment cannot exceed 500 characters'),
    project_link: z.string().url('Invalid project link format').optional(),
    rating: z
        .number()
        .int('Rating must be an integer')
        .min(1, 'Minimum rating is 1')
        .max(5, 'Maximum rating is 5')
        .default(5),
});

export type Feedback = z.infer<typeof FeedbackSchema>;